Added recover job option

+++ GeneContent.pl
@@ -24,6 +24,7 @@ $ProjectName = $ARGV[0];
 $List = $ARGV[1];
 $CPUs = $ARGV[2];
 $MainPath = $ARGV[3];
+$Recover = $ARGV[4];
 
 my($Project, $MainList, $ORFeomesPath, $BlastPath, $ORFsPath,
    $InitialPresenceAbsence, $PresenceAbsence, $PanGenomeSeq, $Stats, $SeqExt,
@@ -71,98 +72,106 @@ open (STDERR, "| tee -ai $LogFile") or die "$0: dup: $!";
 
 @List = ReadFile($MainList);
 $TotalQry = scalar@List;
-@PresenceAbsence = ReadFile($InitialPresenceAbsence);
-$TotalPresenceAbsence = scalar@PresenceAbsence;
-system("rm $InitialPresenceAbsence");
-for ($a=0; $a<$TotalPresenceAbsence; $a++){
-     $Row = $PresenceAbsence[$a];
-     @PresenceAbsenceFields = split(",",$Row);
-     $o = scalar@PresenceAbsenceFields;
-     push (@PresenceAbsenceArray, [@PresenceAbsenceFields]);
-}
-for ($i=0; $i<$TotalPresenceAbsence; $i++){
-        for ($j=0; $j<$o; $j++){
-                if (defined($PresenceAbsenceArray[$i]->[$j])){
-                        $NewReport -> [$i][$j] = $PresenceAbsenceArray[$i]->[$j];
-                }else{
-                        $NewReport -> [$i][$j] = "";
+
+if($Recover == "0"){
+        $Progress = 1;
+        @PresenceAbsence = ReadFile($InitialPresenceAbsence);
+        $TotalPresenceAbsence = scalar@PresenceAbsence;
+        system("rm $InitialPresenceAbsence");
+        for ($a=0; $a<$TotalPresenceAbsence; $a++){
+             $Row = $PresenceAbsence[$a];
+             @PresenceAbsenceFields = split(",",$Row);
+             $o = scalar@PresenceAbsenceFields;
+             push (@PresenceAbsenceArray, [@PresenceAbsenceFields]);
+        }
+        for ($i=0; $i<$TotalPresenceAbsence; $i++){
+                for ($j=0; $j<$o; $j++){
+                        if (defined($PresenceAbsenceArray[$i]->[$j])){
+                                $NewReport -> [$i][$j] = $PresenceAbsenceArray[$i]->[$j];
+                        }else{
+                                $NewReport -> [$i][$j] = "";
+                        }
                 }
         }
+}elsif($Recover == "1"){
+        @Progress = ReadFile($Stats);
+        $Progress = scalar@Progress-2;
 }
-for ($a=1; $a<$TotalQry; $a++){
+        
+for ($a=$Progress; $a<$TotalQry; $a++){
         $QryGenomeName = $List[$a];
         $QryGenomeSeq = $ORFeomesPath ."/". $QryGenomeName . ".ffn";
         $QryDb = $BlastPath ."/". $QryGenomeName;
-        
+                
         $NewReport -> [0][$a+2] = $QryGenomeName;
-        
+                
         @QryIDs = AnnotatedGenes($QryGenomeSeq); #Get all names of annotated genes from query
         $TotalQryIDs = scalar@QryIDs;
         $TotalNewORFs = $TotalQryIDs;
-        
+                
         $nPanGenome = `grep ">" -c $PanGenomeSeq`;
-        
+                
         $Counter = 0;
         @AnalyzedORFs = "";
         for ($b=1; $b<$nPanGenome+1; $b++){
-                
+                        
                 $Counter = sprintf "%.4d", $b;
-                
+                        
                 $TestingORF = "ORF" ."_". $Counter;
                 $ORFpath = $ORFsPath ."/". $TestingORF;
                 $Hmm = $ORFpath ."/". $TestingORF . $HmmExt;
                 $ORFTemp = $ORFpath ."/". $TestingORF ."-". $QryGenomeName . ".temp";
-	
-		print "\n----------------Looking for $TestingORF in $QryGenomeName----------------\n";
-	
-		system("nhmmer -E 1e-10 --cpu $CPUs --noali --dfamtblout $ORFTemp $Hmm $QryGenomeSeq");             
-                
-		@nHMMerReport = ReadFile($ORFTemp);
+        	
+        	print "\n----------------Looking for $TestingORF in $QryGenomeName----------------\n";
+        	
+        	system("nhmmer -E 1e-10 --cpu $CPUs --noali --dfamtblout $ORFTemp $Hmm $QryGenomeSeq");             
+                        
+        	@nHMMerReport = ReadFile($ORFTemp);
                 if(@nHMMerReport){
                         open (FILE, ">$ORFTemp");
                                 print FILE $nHMMerReport[0];
                         close FILE;
                         @nHMMerReport = ReadFile($ORFTemp);
-                        
+                                
                         #$CoreGenomeSize++;
-			$BestHit = $nHMMerReport[0];
-			$BestHit =~ s/\s^//g;
-			$BestHit =~ s/\s+/,/g;
-			@BestHitArray = split(",",$BestHit);
-			$Entry = $BestHitArray[0];
-			$Strand = $BestHitArray[8];
-			$QryORFSeq = $ORFpath ."/". $QryGenomeName ."-". $Entry . $SeqExt;
-                        
+        		$BestHit = $nHMMerReport[0];
+        		$BestHit =~ s/\s^//g;
+        		$BestHit =~ s/\s+/,/g;
+        		@BestHitArray = split(",",$BestHit);
+        		$Entry = $BestHitArray[0];
+        		$Strand = $BestHitArray[8];
+        		$QryORFSeq = $ORFpath ."/". $QryGenomeName ."-". $Entry . $SeqExt;
+                                
                         if ( any { $_ eq $Entry} @AnalyzedORFs){     
                                 print "\n$Entry is already analyzed\n";    
                         }else{
                                 unshift @AnalyzedORFs, $Entry;
-                                
+                                        
                                 $NewReport -> [$b][$a+2] = $Entry;
-                                
+                                        
                                 Extract($QryGenomeName,$QryDb,$Entry,$QryORFSeq);
-                                
+                                        
                                 $PreviousAln = "*-". $TestingORF . $AlnExt;
                                 $FindingPreviousAln = `find $ORFsPath -type f -name $PreviousAln`;
                                 @SplitAlnPath = split("/",$FindingPreviousAln);
                                 $PreviousAlnName = $SplitAlnPath[$#SplitAlnPath];
                                 @SplitPreviousAlnName = split("-",$PreviousAlnName);
                                 $PreviousAlnPrefix = $SplitPreviousAlnName[0];
-                                
+                                        
                                 $CurrentAlnPrefix = $PreviousAlnPrefix+1;
                                 $LastORFAln = $ORFpath ."/". $PreviousAlnPrefix ."-". $TestingORF . $AlnExt;
                                 $NewORFAln = $ORFpath ."/". $CurrentAlnPrefix ."-". $TestingORF . $AlnExt;
-                                
+                                        
                                 print "\tUpdating hmm profile...";
                                 system("hmmalign -o $NewORFAln --trim --mapali $LastORFAln $Hmm $QryORFSeq");
                                 system("hmmbuild $Hmm $NewORFAln");
                                 print "Done!\n";
-                                
+                                        
                                 #Deleting the file obtained from nhmmer 
                                 print "\tCleaning...";
                                 system("rm $ORFTemp $LastORFAln");
                                 print "Done!\n";
-                                
+                                        
                                 #Obataining new ORFs IDs
                                 for($x=0;$x<$TotalNewORFs;$x++){
                                         $ID = $QryIDs[$x];
@@ -172,35 +181,35 @@ for ($a=1; $a<$TotalQry; $a++){
                                         }
                                 }
                         }
-		}else{
+        	}else{
                         print "The ORF $TestingORF is not present in $QryGenomeName\n";
                         $NewReport -> [$b][$a+2] = "";
-                        
+                                
                         system("rm $ORFTemp");
                 }
-	}    
+        }    
         print "\nProcessing New ORFs\n"; # <- Non shared genes from query file
         $TotalNewORFs = scalar@QryIDs;
         for($c=0; $c<$TotalNewORFs; $c++){
                 $NewORFId = $QryIDs[$c];
                 $NewCounter = sprintf "%.4d",$Counter+1+$c;
                 $NewORF = "ORF" ."_". $NewCounter;
-        	$NewORFPath = $ORFsPath ."/". $NewORF;
+                $NewORFPath = $ORFsPath ."/". $NewORF;
                 $NewORFSeq = $NewORFPath ."/". $QryGenomeName ."-". $NewORFId . $SeqExt;
                 $NewORFHmm = $NewORFPath ."/". $NewORF . $HmmExt;
                 $NewORFAln = $NewORFPath ."/". "1-" . $NewORF . $AlnExt;
-        
+                
                 print "\nProcessing ORF $NewCounter: \n";
-        
+                
                 MakeDir($NewORFPath);
                 Extract($QryGenomeName,$QryDb,$NewORFId,$NewORFSeq);
                 $cmd = `cp $NewORFSeq $NewORFAln`;
                 HMM($CPUs,$NewORFHmm,$NewORFAln);
-                
-        	print "\tAdding ORF $NewCounter to PanGenome...";
+                        
+                print "\tAdding ORF $NewCounter to PanGenome...";
                 $cmd = `blastdbcmd -db $QryDb -dbtype nucl -entry "$NewORFId" >> $PanGenomeSeq`;
-        	print "Done!\n";
-                
+                print "Done!\n";
+                        
                 $NewReport -> [$NewCounter][0] = $NewORF;
                 for ($i=1; $i<$a+1;$i++){
                         $NewReport -> [$NewCounter][$i] = "";
@@ -218,34 +227,35 @@ for ($a=1; $a<$TotalQry; $a++){
         
         #Building a file with only those genes that are shared in all genomes (Core-Genome table)
         @SharedORFsArray = ReadFile($PresenceAbsence);
-
+        
         open (FILE,">$CoreGenomeFile");
                 #print FILE "$SharedORFsArray[0]\n";
                 foreach $Lap (@SharedORFsArray){
-                       @LapArray = split(",",$Lap);
-                       chomp@LapArray;
-                       $Count = 0;
-                       foreach $Element(@LapArray){
-                              if ($Element ne ""){
-                                     $Count++;
-                              }
-                       }
-                       if($Count == $a+3){
-                           print FILE "$Lap\n";   
-                       }
+                        @LapArray = split(",",$Lap);
+                        chomp@LapArray;
+                        $Count = 0;
+                        foreach $Element(@LapArray){
+                                if ($Element ne ""){
+                                        $Count++;
+                                }
+                        }
+                        if($Count == $a+3){
+                                print FILE "$Lap\n";   
+                        }
                 }
         close FILE;
-
+        
         @CoreGenome = ReadFile($CoreGenomeFile);
         $q = scalar@CoreGenome;
         $CoreGenes = $q-1;
-        
+                
         $NewStrains = $a+1;
         open (FILE, ">>$Stats");
-	#  "Number Of New Strains,Analyzed Strain,Core Genome,Pan Genome,New Genes\n";
-	print FILE "$NewStrains,$QryGenomeName,$CoreGenes,$NewCounter,$TotalNewORFs\n";
+        #  "Number Of New Strains,Analyzed Strain,Core Genome,Pan Genome,New Genes\n";
+        print FILE "$NewStrains,$QryGenomeName,$CoreGenes,$NewCounter,$TotalNewORFs\n";
         close FILE;
 }
+
 
 open (FILE, ">>$Summary");
         print FILE "Core-Genome Size: $CoreGenes\nPan-Genome Size: $NewCounter";
